MVP-5 Унёс старый экран ввода имени модуля в задепрекейченый пакет
