+++ qpay-module/src/main/kotlin/dev/itssho/module/qpay/module/actor/di/DiCreation.kt
@@ -5,8 +5,8 @@ import dev.itssho.module.qpay.module.actor.di.module.makeCommonDataModule
 import dev.itssho.module.qpay.module.actor.di.module.makeCommonModule
 import dev.itssho.module.qpay.module.actor.di.module.makeCreateDataModule
 import dev.itssho.module.qpay.module.actor.di.module.makeCreateModule
-import dev.itssho.module.qpay.module.actor.di.module.makeNameDataModule
-import dev.itssho.module.qpay.module.actor.di.module.makeNameModule
+import dev.itssho.module.qpay.module.actor.di.module.makeDeprecatedNameDataModule
+import dev.itssho.module.qpay.module.actor.di.module.makeDeprecatedNameModule
 import dev.itssho.module.qpay.module.actor.di.module.makePreparationDataModule
 import dev.itssho.module.qpay.module.actor.di.module.makePreparationModule
 import dev.itssho.module.qpay.module.actor.di.module.makeStructureDataModule
@@ -43,8 +43,8 @@ fun makeDi(jbContext: JBContext): KoinApplication {
 
 
 
-	val nameDataModule = makeNameDataModule().apply { includes(commonDataModule) }
-	val nameModule = makeNameModule().apply {
+	val nameDataModule = makeDeprecatedNameDataModule().apply { includes(commonDataModule) }
+	val nameModule = makeDeprecatedNameModule().apply {
 		includes(nameDataModule)
 		includes(commonModule)
 	}
