+++ qpay-module/src/main/kotlin/dev/itssho/module/qpay/module/actor/di/module/DeprecatedNameModule.kt
@@ -2,9 +2,9 @@
 
 package dev.itssho.module.qpay.module.actor.di.module
 
-import dev.itssho.module.qpay.module.actor.di.component.QpayNameKoinDi
-import dev.itssho.module.qpay.module.name.presentation.QpayNameViewModel
-import dev.itssho.module.qpay.module.name.ui.NameUI
+import dev.itssho.module.qpay.module.actor.di.component.QpayDeprecatedNameKoinDi
+import dev.itssho.module.qpay.module.name.deprecated.presentation.QpayNameViewModel
+import dev.itssho.module.qpay.module.name.deprecated.ui.NameUI
 import dev.itssho.module.qpay.module.structure.actor.di.UiScopeQ
 import dev.itssho.module.util.koin.factoryScopeOf
 import kotlinx.coroutines.CoroutineScope
@@ -14,10 +14,10 @@ import kotlinx.coroutines.swing.Swing
 import org.koin.core.module.dsl.scopedOf
 import org.koin.dsl.module
 
-fun makeNameDataModule() = module {}
+fun makeDeprecatedNameDataModule() = module {}
 
-fun makeNameModule() = module {
-	factoryScopeOf(::QpayNameKoinDi) {
+fun makeDeprecatedNameModule() = module {
+	factoryScopeOf(::QpayDeprecatedNameKoinDi) {
 		scoped(UiScopeQ) { CoroutineScope(Job() + Dispatchers.Swing) }
 		scopedOf(::QpayNameViewModel)
 		scoped { NameUI(get(), get(), get<CoroutineScope>(UiScopeQ)) }
