+++ qpay-module/src/main/kotlin/dev/itssho/module/qpay/module/name/deprecated/presentation/QpayNameViewModel.kt
@@ -1,20 +1,20 @@
-package dev.itssho.module.qpay.module.name.presentation
+package dev.itssho.module.qpay.module.name.deprecated.presentation
 
 import dev.itssho.module.component.chain.Separator
 import dev.itssho.module.component.chain.chainOf
 import dev.itssho.module.component.chain.splitToChain
 import dev.itssho.module.core.presentation.ViewModel
 import dev.itssho.module.qpay.module.common.domain.usecase.SetModuleNameUseCase
-import dev.itssho.module.qpay.module.name.domain.entity.ModuleDomain
-import dev.itssho.module.qpay.module.name.domain.entity.ModuleLevel
-import dev.itssho.module.qpay.module.name.presentation.converter.convertDomainName
-import dev.itssho.module.qpay.module.name.presentation.converter.convertLevelName
-import dev.itssho.module.qpay.module.name.presentation.converter.convertModuleName
-import dev.itssho.module.qpay.module.name.presentation.converter.parseModuleName
-import dev.itssho.module.qpay.module.name.presentation.model.ModuleNameValidationResult
-import dev.itssho.module.qpay.module.name.presentation.model.ModuleNameValidationResult.Valid
-import dev.itssho.module.qpay.module.name.presentation.state.ModuleNameState
-import dev.itssho.module.qpay.module.name.presentation.validator.validateModuleName
+import dev.itssho.module.qpay.module.name.deprecated.domain.entity.ModuleDomain
+import dev.itssho.module.qpay.module.name.deprecated.domain.entity.ModuleLevel
+import dev.itssho.module.qpay.module.name.deprecated.presentation.converter.convertDomainName
+import dev.itssho.module.qpay.module.name.deprecated.presentation.converter.convertLevelName
+import dev.itssho.module.qpay.module.name.deprecated.presentation.converter.convertModuleName
+import dev.itssho.module.qpay.module.name.deprecated.presentation.converter.parseModuleName
+import dev.itssho.module.qpay.module.name.deprecated.presentation.model.ModuleNameValidationResult
+import dev.itssho.module.qpay.module.name.deprecated.presentation.model.ModuleNameValidationResult.Valid
+import dev.itssho.module.qpay.module.name.deprecated.presentation.state.ModuleNameState
+import dev.itssho.module.qpay.module.name.deprecated.presentation.validator.validateModuleName
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import java.util.concurrent.atomic.AtomicBoolean
