+++ qpay-module/src/main/kotlin/dev/itssho/module/qpay/module/name/deprecated/presentation/model/ModuleNameValidationResult.kt
@@ -1,4 +1,4 @@
-package dev.itssho.module.qpay.module.name.presentation.model
+package dev.itssho.module.qpay.module.name.deprecated.presentation.model
 
 sealed interface ModuleNameValidationResult {
